feat: implemented ConfigDecorator
