+++ src/components/DatePicker/index.tsx
@@ -1,33 +1,8 @@
-import { useMemo, useState } from 'react';
-
-import { DatePickerProvider } from '@/components/DatePickerProvider';
-import { SimpleDatePicker } from '@/components/SimpleDatePicker';
-import { ActiveDayContext } from '@/context';
+import { ConfigDecorator } from '@/services';
 import { DatePickerConfigType } from '@/types';
 
-const INITIAL_ACTIVE_DAY_VALUE = null;
+export const DatePicker = (config: DatePickerConfigType) => {
+  const { DatePickerComponent } = new ConfigDecorator(config);
 
-interface DatePickerProps extends DatePickerConfigType {
-  initialDate?: Date;
-}
-
-export const DatePicker = ({ initialDate, ...config }: DatePickerProps) => {
-  const [activeDay, setActiveDay] = useState<number | null>(
-    initialDate?.getTime() ?? INITIAL_ACTIVE_DAY_VALUE,
-  );
-
-  const resetActiveDay = () => setActiveDay(INITIAL_ACTIVE_DAY_VALUE);
-
-  const activeDayContextObj = useMemo(
-    () => ({ activeDay, setActiveDay, resetActiveDay }),
-    [activeDay],
-  );
-
-  return (
-    <DatePickerProvider config={config}>
-      <ActiveDayContext.Provider value={activeDayContextObj}>
-        <SimpleDatePicker />
-      </ActiveDayContext.Provider>
-    </DatePickerProvider>
-  );
+  return <DatePickerComponent />;
 };
