--- src/components/RangeDatePicker/styled.ts
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const Container = styled.div`
-  position: relative;
-
-  max-width: ${({ theme }) => theme.datePickerMaxWidth};
-`;
