+++ src/components/index.ts
@@ -1,2 +1,1 @@
 export { DatePicker } from './DatePicker';
-export { RangeDatePicker } from './RangeDatePicker';
