+++ src/hocs/index.ts
@@ -1,3 +1,6 @@
+export { withProviders } from './withProviders';
 export { withRangeCalendar } from './withRangeCalendar';
+export { withRangeLogic } from './withRangeLogic';
 export { withSimpleCalendar } from './withSimpleCalendar';
+export { withSimpleLogic } from './withSimpleLogic';
 export { withTodos } from './withTodos';
