+++ src/hooks/index.ts
@@ -1,5 +1,6 @@
 export { useCalendar } from './useCalendar';
 export { useCalendarCell } from './useCalendarCell';
+export { useCalendarToggle } from './useCalendarToggle';
 export { useDateInput } from './useDateInput';
 export { useRangeDateInput } from './useRangeDateInput';
 export { useTodos } from './useTodos';
