+++ src/services/index.ts
@@ -1,3 +1,4 @@
+export { ConfigDecorator } from './ConfigDecorator';
 export { DayFactory } from './DayFactory';
 export { holidays } from './HolidaysDataBase';
 export { TodosControlDecorator } from './TodosControlDecorator';
