+++ src/styles/index.ts
@@ -1,3 +1,3 @@
-export { GlobalStyles } from './GlobalStyles';
+export * from './base';
 export * from './mixins';
 export { theme } from './theme';
