+++ src/types/DatePickerConfigType.ts
@@ -3,6 +3,10 @@ import { CalendarView, CalendarWeekStart } from '@/constants';
 import { ThemeType } from './ThemeType';
 
 export interface DatePickerConfigType {
+  range?: boolean;
+  initialDate?: Date;
+  initialStartDate?: Date;
+  initialEndDate?: Date;
   label?: string;
   showHolidays?: boolean;
   showWeekends?: boolean;
